refactor: isolate compression summary dispatch
